[#9] Set directory structure
- [x] Add tslint
- [x] Divide codes into 'main' and 'test'
- [x] Replace css files with JSS

- TODO: Add build, run script
