+++ guestbook-webapp/package.json
@@ -14,13 +14,18 @@
     "@types/react": "16.8.16",
     "@types/react-dom": "16.8.4",
     "@types/react-jss": "^8.6.3",
-    "typescript": "3.4.5"
+    "typescript": "3.4.5",
+    "tslint": "^5.16.0",
+    "tslint-config-prettier": "^1.18.0",
+    "tslint-react": "^4.0.0",
+    "tslint-sonarts": "^1.9.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint": "tslint -c tslint.json -p tsconfig.json"
   },
   "eslintConfig": {
     "extends": "react-app"
