+++ guestbook-webapp/src/index.tsx
@@ -1,8 +1,29 @@
+import Home from 'main/ui/component/templates/Home';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import withStyles from 'react-jss';
 import * as serviceWorker from './serviceWorker';
+
+const styles = {
+  '@global': {
+    body: {
+      margin: 0,
+      padding: 0,
+      fontFamily: `-apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif`,
+      "-webkit-font-smoothing": "antialiased",
+      "-moz-osx-font-smoothing": "grayscale"
+    },
+
+    code: {
+      fontFamily: `source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace`
+    }
+  }
+}
+
+const App = withStyles(styles)(() => <Home />)
 
 ReactDOM.render(<App />, document.getElementById('root'));
