+++ guestbook-webapp/src/test/ui/component/templates/App.test.tsx
@@ -1,9 +1,9 @@
+import Home from 'main/ui/component/templates/Home';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  ReactDOM.render(<Home />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
